Attempt at a test for bugzilla 1950

+++ unixConfig/CollateCore/TargetApp/dumpcore.py
@@ -1,11 +1,15 @@
 #!/usr/bin/env python
 
-import os, time
+import os, time, sys
 
-processId = os.fork()
-if processId == 0:
-    time.sleep(10)
+if len(sys.argv) == 1:
+    processId = os.fork()
+    if processId == 0:
+        time.sleep(10)
+    else:
+        time.sleep(1)
+        print "Killing child process with signal 11"
+        os.system("kill -11 " + str(processId))
 else:
-    time.sleep(1)
-    print "Killing child process with signal 11"
-    os.system("kill -11 " + str(processId))
+    # Produce fake core
+    os.system("which python > core")
